+++ source/extendAlign.cpp
@@ -23,7 +23,7 @@ if (extendToEnd) {//end to end extension
         iS=dR*iExt;
         iG=dG*iExt;
 
-        if ((gStart+iG)==-1 || G[iG]==5) {//prohibit extension through chr boundary
+        if ((gStart+iG)==(uint)(-1) || G[iG]==5) {//prohibit extension through chr boundary
             trA->extendL=0;
             trA->maxScore=-999999999;
             trA->nMatch=0;
@@ -61,7 +61,7 @@ for (int i=0;i<(int) L;i++) {
     iS=dR*i;
     iG=dG*i;
     
-    if ((gStart+iG)==-1 || G[iG]==5 || R[iS]==MARK_FRAG_SPACER_BASE) break; //no extension through chr boundary, or through the spacer between fragments
+    if ((gStart+iG)==(uint)(-1) || G[iG]==5 || R[iS]==MARK_FRAG_SPACER_BASE) break; //no extension through chr boundary, or through the spacer between fragments
     if (R[iS]>3 || G[iG]>3) continue;//no penalties for Ns in reads or genome
     
     if (G[iG]==R[iS]) {//Match
